+++ miheev.egor/S1/mainUtils.hpp
@@ -3,17 +3,17 @@
 
 #include "list.hpp"
 
-using SI_pair = std::pair< std::string, miheev::List< int > >;
-using ListIter = miheev::List<int>::Iterator;
+using SI_pair = std::pair< std::string, miheev::List< size_t > >;
+using ListIter = miheev::List<size_t>::Iterator;
 
 namespace miheev
 {
   void expandPairsArr(SI_pair*& pairs, size_t& size);
-  List< int > readNumbers(std::istream& stream);
+  List< size_t > readNumbers(std::istream& stream);
   SI_pair getSIPair(std::istream& stream);
   ListIter* getIters(SI_pair* pairs, size_t size);
-  List< int >* getLists(SI_pair* pairs, size_t size);
-  size_t maxListSize(List< int >* lists, size_t size);
+  List< size_t >* getLists(SI_pair* pairs, size_t size);
+  size_t maxListSize(List< size_t >* lists, size_t size);
 }
 
 #endif
